+++ Sources/Sentry/SentrySerialization.m
@@ -7,6 +7,7 @@
 #import "SentryError.h"
 #import "SentryLevelMapper.h"
 #import "SentryLogC.h"
+#import "SentryModels+Serializable.h"
 #import "SentrySdkInfo.h"
 #import "SentrySession.h"
 #import "SentrySwift.h"