+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-# Unreleased
+## Unreleased
+
+### Features
+
+- Add experimental support for capturing structured logs via `SentrySDK.logger` (#5532)
 
 ### Improvements
 
@@ -11,6 +15,9 @@
 ### Fixes
 
 - Fix video replay crashes due to video writer inputs not marked as finished on cancellation (#5608)
+- Fix wrong flush timeout (#5565). When flush timed out before the SDK finished sending data, it always blocked the full flush timeout the next time being called. This is fixed now.
+
+- Launch profiling now respects original configured options if they change on the next launch (#5417)
 
 ## 8.53.2